Add checks/workaround to potential chat download issues
During additional testing, found some corner cases where TDCLI fails to download any chat logs.

One of which is missing or malformed emojis that we try to embed. If this occurs, we'll try a second time without downloading/embedding emojies.

Worst case if it still fails, we'll assume a chat log does not exist and write a placeholder json file to satisfy future file checks.

Issue documented in lay295/TwitchDownloader#396 - will revert/modify this code if an official fix is implemented.