Migrated ffmpeg to a pip dependency and TDCLI calls to config.yaml
Tightened up the bolts and streamlined some things.

Moved the ffmpeg dependency to the pip static-ffmpeg library. architecture agnostic and we no longer rely on the user to download, set correct paths, or needing to store ffmpeg binaries in the repo.

Also cleaned up the TwitchDownloaderCLI calls and moved the path out to a renamed config.yaml